// src/app/shared/model/verf-question-grp-d-obj.model.ts
export class VerfQuestionGrpDObj {
    VerfQuestionGrpDId: number;
    VerfQuestionGrpHId: number;
    VerfQuestionAnswerId: number;
    SeqNo: number;
    IsActive: boolean;
    ListVerfQuestionAnswerId: number[];
    RowVersion : string;
    constructor() { this.VerfQuestionAnswerId = 0; this.VerfQuestionGrpHId = 0; this.VerfQuestionGrpDId = 0 }
}  